+++ compiler-core/src/analyse/imports.rs
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2023 The Gleam contributors
+
 use ecow::EcoString;
 
 use crate::{
@@ -295,12 +298,15 @@ impl<'context, 'problems> Importer<'context, 'problems> {
                 import.module.clone(),
                 alias_location,
                 import.location,
+                import.module_location,
             );
         } else {
             self.environment.references.register_module(
                 used_name.clone(),
                 import.module.clone(),
                 import.location,
+                import.module_location,
+                None,
             );
         }