+++ compiler-core/src/javascript/tests.rs
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2021 The Gleam contributors
+
 use crate::{
     analyse::TargetSupport,
     build::{Origin, Target},
@@ -173,7 +176,7 @@ pub fn compile(src: &str, deps: Vec<(&str, &str, &str)>) -> TypedModule {
             importable_modules: &modules,
             warnings: &TypeWarningEmitter::null(),
             direct_dependencies: &HashMap::new(),
-            dev_dependencies: &std::collections::HashSet::new(),
+            dev_dependencies: &HashSet::new(),
             target_support: TargetSupport::Enforced,
             package_config: &dep_config,
         }
@@ -199,7 +202,7 @@ pub fn compile(src: &str, deps: Vec<(&str, &str, &str)>) -> TypedModule {
         importable_modules: &modules,
         warnings: &TypeWarningEmitter::null(),
         direct_dependencies: &direct_dependencies,
-        dev_dependencies: &std::collections::HashSet::new(),
+        dev_dependencies: &HashSet::new(),
         target_support: TargetSupport::NotEnforced,
         package_config: &config,
     }