+++ compiler-core/src/javascript/tests/custom_types.rs
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2021 The Gleam contributors
+
 use crate::javascript::tests::CURRENT_PACKAGE;
 use crate::{assert_js, assert_ts_def};
 
@@ -1068,5 +1071,123 @@ pub type Box {
   Box(constructor: Int)
 }
 "#,
+    );
+}
+
+#[test]
+fn singleton_for_two_aliased_variants() {
+    assert_js!(
+        ("wibble", "pub type Wibble { Wibble }"),
+        ("wobble", "pub type Wibble { Wibble }"),
+        "
+import wobble.{Wibble as Wobble}
+import wibble.{Wibble as Wubble}
+
+pub fn main() {
+  #(Wobble, Wubble)
+}
+"
+    );
+}
+
+#[test]
+fn fieldless_variant_still_imported_if_used_in_instanceof_comparison() {
+    assert_js!(
+        ("wibble", "pub type Wibble { Wibble Wobble }"),
+        "
+import wibble.{Wibble}
+
+pub fn main() {
+  let x = Wibble
+  x == Wibble
+}
+"
+    );
+}
+
+#[test]
+fn fieldless_variant_still_imported_if_used_in_instanceof_comparison_in_guard() {
+    assert_js!(
+        ("wibble", "pub type Wibble { Wibble Wobble }"),
+        "
+import wibble.{Wibble}
+
+pub fn main() {
+  let x = Wibble
+  case x {
+    _ if x == Wibble -> 1
+    _ -> 2
+  }
+}
+"
+    );
+}
+
+#[test]
+fn aliased_variant_still_imported_if_used_in_instanceof_comparison() {
+    assert_js!(
+        ("wibble", "pub type Wibble { Wibble Wobble }"),
+        "
+import wibble.{Wibble as Wobble}
+
+pub fn main() {
+  let x = Wobble
+  x == Wobble
+}
+"
+    );
+}
+
+#[test]
+fn aliased_variant_still_imported_if_used_in_instanceof_comparison_in_guard() {
+    assert_js!(
+        ("wibble", "pub type Wibble { Wibble Wobble }"),
+        "
+import wibble.{Wibble as Wobble}
+
+pub fn go(x) {
+  case x {
+    _ if x == Wobble -> Wobble
+    _ -> wibble.Wobble
+  }
+}
+"
+    );
+}
+
+#[test]
+fn aliased_variant_still_imported_if_used_in_pattern_matching() {
+    assert_js!(
+        ("wibble", "pub type Wibble { Wibble Wobble }"),
+        "
+import wibble.{Wibble as Wobble}
+
+pub fn go(x) {
+  case x {
+    Wobble -> wibble.Wobble
+    _ -> Wobble
+  }
+}
+"
+    );
+}
+
+#[test]
+fn opaque_variant_produces_private_singleton_constants() {
+    assert_js!(
+        "
+pub opaque type Wibble {
+  Wibble
+  Wobble(Int)
+}
+
+pub fn wibble() -> Wibble {
+  Wibble
+}
+
+pub fn wobble(x: Int) -> Wibble {
+  Wobble(x)
+}
+"
     );
 }