+++ compiler-core/src/parse.rs
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2020 The Gleam contributors
+
 // Gleam Parser
 //
 // Terminology:
@@ -3003,7 +3006,6 @@ where
         let mut start = 0;
         let mut end;
         let mut module = EcoString::new();
-        let mut last_segment_start;
         let mut last_segment_end;
 
         // Gather module names
@@ -3016,7 +3018,6 @@ where
             }
             module.push_str(&name);
             end = e;
-            last_segment_start = s;
             last_segment_end = e;
 
             // Useful error for : import a/.{b}
@@ -3097,7 +3098,7 @@ where
                 end,
             },
             module_location: SrcSpan {
-                start: last_segment_start,
+                start,
                 end: last_segment_end,
             },
             unqualified_values,