+++ language-server/src/tests.rs
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2023 The Gleam contributors
+
 mod action;
 mod compilation;
 mod completion;
@@ -805,6 +808,20 @@ impl<'a> TestProject<'a> {
             .expect("Module doesn't exist");
 
         executor(&mut engine, params, code.into())
+    }
+
+    /// Run a test in a project without a specific position (for workspace-wide
+    /// actions).
+    pub fn run<T>(
+        &self,
+        executor: impl FnOnce(
+            &mut LanguageServerEngine<LanguageServerTestIO, LanguageServerTestIO>,
+        ) -> T,
+    ) -> T {
+        // Use a throwaway position and ignore it
+        let (mut engine, _) = self.positioned_with_io(Position::default());
+
+        executor(&mut engine)
     }
 }